Complete AWS setup and modify xml to configure package for export.
AWS encountering errors not seen locally; require more troubleshooting/debugging.  Expectation is DB connection/c3p0 issue, but uncertain.  Tests continue.

+++ src/main/java/trekroulette/controller/RouletteController.java
@@ -14,7 +14,7 @@ public class RouletteController {
     }
 
     public void getMaximumEpisodeId(List<Episode> library) {
-        maximumEpisodeId = library.size();
+        maximumEpisodeId = (int) library.size();
     }
 
     public void getRandomEpisodeId() {
